+++ block-core/src/lib.rs
@@ -1,0 +1,17 @@
+#![cfg_attr(not(feature = "std"), no_std)]
+#![cfg_attr(not(feature = "std"), feature(alloc))]
+
+#[cfg(not(feature = "std"))]
+extern crate alloc;
+
+extern crate sha3;
+extern crate rlp;
+extern crate bigint;
+
+mod account;
+mod log;
+mod transaction;
+
+pub use transaction::*;
+pub use account::Account;
+pub use log::Log;
