+++ block-core/src/log.rs
@@ -1,6 +1,9 @@
 use rlp::{Encodable, Decodable, RlpStream, DecoderError, UntrustedRlp};
 use bigint::{Address, Gas, H256, U256, B256, H64};
 
+#[cfg(not(feature = "std"))]
+use alloc::Vec;
+
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub struct Log {
     pub address: Address,
