+++ block/src/lib.rs
@@ -5,23 +5,21 @@ extern crate secp256k1;
 extern crate sha3;
 extern crate blockchain;
 extern crate trie;
+extern crate block_core;
 #[cfg(test)] extern crate hexutil;
 #[cfg(test)] extern crate rand;
 
 mod header;
 mod transaction;
 mod block;
-mod account;
 mod receipt;
-mod log;
 mod address;
 
+pub use block_core::*;
 pub use transaction::*;
 pub use header::{TotalHeader, Header};
 pub use block::{Block, transactions_root, receipts_root, ommers_hash};
-pub use account::Account;
 pub use receipt::Receipt;
-pub use log::Log;
 pub use address::FromKey;
 
 use bigint::H256;
